Break down form view component into smaller components
